+++ kurtosis-module/impl/participant_network/participant_network.go
@@ -52,7 +52,7 @@ func LaunchParticipantNetwork(
 	prelaunchDataGeneratorCtx *prelaunch_data_generator.PrelaunchDataGeneratorContext,
 	networkParams *module_io.NetworkParams,
 	allParticipantSpecs []*module_io.ParticipantParams,
-	logLevel module_io.ParticipantLogLevel,
+	globalLogLevel module_io.GlobalClientLogLevel,
 	shouldWaitForMining bool,
 ) (
 	resultParticipants []*Participant,
@@ -134,8 +134,10 @@ func LaunchParticipantNetwork(
 				enclaveCtx,
 				elClientServiceId,
 				participantSpec.ELClientImage,
-				logLevel,
+				participantSpec.ELClientLogLevel,
+				globalLogLevel,
 				elClientContextForBootElClients,
+				participantSpec.ELExtraParams,
 			)
 		} else {
 			bootElClientCtx := allElClientContexts[bootParticipantIndex]
@@ -143,8 +145,10 @@ func LaunchParticipantNetwork(
 				enclaveCtx,
 				elClientServiceId,
 				participantSpec.ELClientImage,
-				logLevel,
+				participantSpec.ELClientLogLevel,
+				globalLogLevel,
 				bootElClientCtx,
+				participantSpec.ELExtraParams,
 			)
 		}
 		if elClientLaunchErr != nil {
@@ -247,10 +251,13 @@ func LaunchParticipantNetwork(
 				enclaveCtx,
 				clClientServiceId,
 				participantSpec.CLClientImage,
-				logLevel,
+				participantSpec.CLClientLogLevel,
+				globalLogLevel,
 				clClientContextForBootClClients,
 				elClientCtx,
 				newClNodeValidatorKeystores,
+				participantSpec.BeaconExtraParams,
+				participantSpec.ValidatorExtraParams,
 			)
 		} else {
 			bootClClientCtx := allClClientContexts[bootParticipantIndex]
@@ -258,10 +265,13 @@ func LaunchParticipantNetwork(
 				enclaveCtx,
 				clClientServiceId,
 				participantSpec.CLClientImage,
-				logLevel,
+				participantSpec.CLClientLogLevel,
+				globalLogLevel,
 				bootClClientCtx,
 				elClientCtx,
 				newClNodeValidatorKeystores,
+				participantSpec.BeaconExtraParams,
+				participantSpec.ValidatorExtraParams,
 			)
 		}
 		if clClientLaunchErr != nil {