+++ docs/changelog.md
@@ -1,5 +1,19 @@
 # TBD
 
+# 0.4.3
+### Features
+* Add extra debug logging to EL REST client, for debugging any issues
+* Add new module's params `elLogLevel` and `clLogLevel` to configure a specific EL and CL client's log level
+* Added the `elExtraParams`, `beaconExtraParams`, and `validatorExtraParams` properties to a participant to allow for overriding participant commands
+
+### Changes
+* Set config values to `BELLATRIX_` rather than `MERGE_`
+* Set `ETH1_FOLLOW_DISTANCE` to 30
+
+### Fixes
+* Use emptystring for Besu ENR, as there's no way to get it right now without the logs
+* Fixed a Teku break caused by a flag getting renamed in the latest version of the Teku image
+
 # 0.4.2
 ### Features
 * Added generation of Besu genesis file